@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the jsonline library.  If not, see <http://www.gnu.org/licenses/>.
 
-//nolint:dupl
+//nolint:dupl,varnamelen
 package jsonline_test
 
 import (
